Game 2048 (1st variant) : the base cannot be lower than 2

+++ src/game_2048_num.hpp
@@ -2,16 +2,18 @@
 #define GAME_2048_NUM_HPP
 
 #include "game_2048.hpp"
+#include <cmath>
 
 class Game_2048_Num : public Game_2048 {
 	/* Variant with other numbers than powers of 2 */
 public:
-	Game_2048_Num(int height, long base=2);
+	Game_2048_Num(int height, int base = 2);
 
 private:
 	const long base;
 	
 	virtual long random_value();
 };
+
 
 #endif
